+++ era-compiler-solidity/src/linker/output/mod.rs
@@ -2,12 +2,14 @@
 //! The linker output.
 //!
 
-pub mod contract;
+pub mod ignored;
+pub mod linked;
 pub mod unlinked;
 
 use std::collections::BTreeMap;
 
-use self::contract::Contract;
+use self::ignored::Ignored;
+use self::linked::Linked;
 use self::unlinked::Unlinked;
 
 ///
@@ -16,9 +18,9 @@ use self::unlinked::Unlinked;
 #[derive(Debug, Default, serde::Serialize)]
 pub struct Output {
     /// Linked bytecode files with bytecode hashes.
-    pub linked: BTreeMap<String, Contract>,
+    pub linked: BTreeMap<String, Linked>,
     /// Unlinked bytecode files with the list of unlinked symbols.
     pub unlinked: BTreeMap<String, Unlinked>,
     /// Ignored bytecode files that do not require linking.
-    pub ignored: BTreeMap<String, Contract>,
+    pub ignored: BTreeMap<String, Ignored>,
 }